# Apply min_p to unsorted tokens