@JohannesGaessler 2024-01-25
Unless I'm misunderstanding something the software design is that those sampling functions that need probabilities call `llama_sample_softmax` at the beginning to ensure that they are defined. In particular, when after `sampler_queue` `llama_sample_token` is called there is a call to `llama_sample_softmax`.