@ikawrakow 2024-01-25
> Unless I'm misunderstanding something the software design is that those sampling functions that need probabilities call `llama_sample_softmax` at the beginning to ensure that they are defined. In particular, when after `sampler_queue` `llama_sample_token` is called there is a call to `llama_sample_softmax`.

Yes, and because the original version of `llama_sample_min_p` calls `llama_sample_softmax` as the very first thing, someone somewhere could have assumed they don't need to do that after having called `llama_sample_min_p`. No? I'm not saying this is the case, but asking to make sure that this is not the case.