@ggerganov 2024-01-26
Please consider adding some tests for these sampling changes - we have some in `tests/test-sampling.cpp`, but looks like the bare minimum of what we need. Would make it easier to review such kind of changes