@ggerganov 2024-01-28
I think it is okay to remove the softmax from `min_p` since the API does not indicate that softmax would by applied when calling `llama_sample_min_p`. It would be an incorrect assumption if someone relies on having probabilities after `llama_sample_min_p`