@JohannesGaessler 2024-01-28
I rebased this PR onto master. The results for the tests added in https://github.com/ggerganov/llama.cpp/pull/5147 do not change.